+++ src/bin/afl_runner.rs
@@ -1,8 +1,9 @@
-fn main() {
-    extern crate afl;
-    extern crate cpp_demangle;
+#[macro_use]
+extern crate afl;
+extern crate cpp_demangle;
 
-    afl::read_stdio_bytes(|bytes| {
+fn main() {
+    afl::fuzz!(|bytes| {
         let _ = cpp_demangle::Symbol::new(bytes);
     });
 }