+++ src/lib.rs
@@ -30,12 +30,10 @@
 #![deny(missing_docs)]
 #![deny(missing_debug_implementations)]
 #![deny(unsafe_code)]
-
 // Clippy stuff.
 #![allow(unknown_lints)]
 #![allow(clippy::inline_always)]
 #![allow(clippy::redundant_field_names)]
-
 #![cfg_attr(all(not(feature = "std"), feature = "alloc"), no_std)]
 #![cfg_attr(all(not(feature = "std"), feature = "alloc"), feature(alloc))]
 
@@ -82,6 +80,26 @@ use ast::{Demangle, Parse, ParseContext};
 use error::{Error, Result};
 use index_str::IndexStr;
 use std::fmt;
+use std::num::NonZeroU32;
+
+/// Options to control the parsing process.
+#[derive(Clone, Copy, Debug, Default)]
+#[repr(C)]
+pub struct ParseOptions {
+    recursion_limit: Option<NonZeroU32>,
+}
+
+impl ParseOptions {
+    /// Set the limit on recursion depth during the parsing phase. A low
+    /// limit will cause valid symbols to be rejected, but a high limit may
+    /// allow pathological symbols to overflow the stack during parsing.
+    /// The default value is 96, which will not overflow the stack even in
+    /// a debug build.
+    pub fn recursion_limit(mut self, limit: u32) -> Self {
+        self.recursion_limit = Some(NonZeroU32::new(limit).expect("Recursion limit must be > 0"));
+        self
+    }
+}
 
 /// Options to control the demangling process.
 #[derive(Clone, Copy, Debug, Default)]
@@ -89,6 +107,7 @@ use std::fmt;
 pub struct DemangleOptions {
     no_params: bool,
     no_return_type: bool,
+    recursion_limit: Option<NonZeroU32>,
 }
 
 impl DemangleOptions {
@@ -106,6 +125,15 @@ impl DemangleOptions {
     /// Do not display the function return type.
     pub fn no_return_type(mut self) -> Self {
         self.no_return_type = true;
+        self
+    }
+
+    /// Set the limit on recursion depth during the demangling phase. A low
+    /// limit will cause valid symbols to be rejected, but a high limit may
+    /// allow pathological symbols to overflow the stack during demangling.
+    /// The default value is 128.
+    pub fn recursion_limit(mut self, limit: u32) -> Self {
+        self.recursion_limit = Some(NonZeroU32::new(limit).expect("Recursion limit must be > 0"));
         self
     }
 }
@@ -131,7 +159,8 @@ impl<T> Symbol<T>
 where
     T: AsRef<[u8]>,
 {
-    /// Given some raw storage, parse the mangled symbol from it.
+    /// Given some raw storage, parse the mangled symbol from it with the default
+    /// options.
     ///
     /// ```
     /// use cpp_demangle::Symbol;
@@ -161,11 +190,49 @@ where
     ///     "JS_GetPropertyDescriptorById(JSContext*, JS::Handle<JSObject*>, JS::Handle<jsid>, JS::MutableHandle<JS::PropertyDescriptor>)"
     /// );
     /// ```
+    #[inline]
     pub fn new(raw: T) -> Result<Symbol<T>> {
+        Self::new_with_options(raw, &Default::default())
+    }
+
+    /// Given some raw storage, parse the mangled symbol from it.
+    ///
+    /// ```
+    /// use cpp_demangle::{ParseOptions, Symbol};
+    /// use std::string::ToString;
+    ///
+    /// // First, something easy :)
+    ///
+    /// let mangled = b"_ZN5space3fooEibc";
+    ///
+    /// let parse_options = ParseOptions::default()
+    ///     .recursion_limit(1024);
+    ///
+    /// let sym = Symbol::new_with_options(&mangled[..], &parse_options)
+    ///     .expect("Could not parse mangled symbol!");
+    ///
+    /// let demangled = sym.to_string();
+    /// assert_eq!(demangled, "space::foo(int, bool, char)");
+    ///
+    /// // Now let's try something a little more complicated!
+    ///
+    /// let mangled =
+    ///     b"__Z28JS_GetPropertyDescriptorByIdP9JSContextN2JS6HandleIP8JSObjectEENS2_I4jsidEENS1_13MutableHandleINS1_18PropertyDescriptorEEE";
+    ///
+    /// let sym = Symbol::new(&mangled[..])
+    ///     .expect("Could not parse mangled symbol!");
+    ///
+    /// let demangled = sym.to_string();
+    /// assert_eq!(
+    ///     demangled,
+    ///     "JS_GetPropertyDescriptorById(JSContext*, JS::Handle<JSObject*>, JS::Handle<jsid>, JS::MutableHandle<JS::PropertyDescriptor>)"
+    /// );
+    /// ```
+    pub fn new_with_options(raw: T, options: &ParseOptions) -> Result<Symbol<T>> {
         let mut substitutions = subs::SubstitutionTable::new();
 
         let parsed = {
-            let ctx = ParseContext::default();
+            let ctx = ParseContext::new(*options);
             let input = IndexStr::new(raw.as_ref());
 
             let (parsed, tail) = ast::MangledName::parse(&ctx, &mut substitutions, input)?;
@@ -236,13 +303,13 @@ substitutions = {:#?}",
     /// Demangle the symbol to a DemangleWrite, which lets the consumer be informed about
     /// syntactic structure.
     #[allow(clippy::trivially_copy_pass_by_ref)]
-    pub fn structured_demangle<W: DemangleWrite>(&self, out: &mut W, options: &DemangleOptions) -> fmt::Result {
-        let mut ctx = ast::DemangleContext::new(
-            &self.substitutions,
-            self.raw.as_ref(),
-            *options,
-            out,
-        );
+    pub fn structured_demangle<W: DemangleWrite>(
+        &self,
+        out: &mut W,
+        options: &DemangleOptions,
+    ) -> fmt::Result {
+        let mut ctx =
+            ast::DemangleContext::new(&self.substitutions, self.raw.as_ref(), *options, out);
         self.parsed.demangle(&mut ctx, None)
     }
 }
@@ -292,10 +359,11 @@ impl<W: fmt::Write> DemangleWrite for W {
 }
 
 impl<'a, T> Symbol<&'a T>
-    where T: AsRef<[u8]> + ?Sized,
+where
+    T: AsRef<[u8]> + ?Sized,
 {
     /// Parse a mangled symbol from input and return it and the trailing tail of
-    /// bytes that come after the symbol.
+    /// bytes that come after the symbol, with the default options.
     ///
     /// While `Symbol::new` will return an error if there is unexpected trailing
     /// bytes, `with_tail` simply returns the trailing bytes along with the
@@ -315,10 +383,42 @@ impl<'a, T> Symbol<&'a T>
     /// let demangled = sym.to_string();
     /// assert_eq!(demangled, "space::foo(int, bool, char)");
     /// ```
+    #[inline]
     pub fn with_tail(input: &'a T) -> Result<(BorrowedSymbol<'a>, &'a [u8])> {
+        Self::with_tail_and_options(input, &Default::default())
+    }
+
+    /// Parse a mangled symbol from input and return it and the trailing tail of
+    /// bytes that come after the symbol.
+    ///
+    /// While `Symbol::new_with_options` will return an error if there is
+    /// unexpected trailing bytes, `with_tail_and_options` simply returns the
+    /// trailing bytes along with the parsed symbol.
+    ///
+    /// ```
+    /// use cpp_demangle::{BorrowedSymbol, ParseOptions};
+    /// use std::string::ToString;
+    ///
+    /// let mangled = b"_ZN5space3fooEibc and some trailing junk";
+    ///
+    /// let parse_options = ParseOptions::default()
+    ///     .recursion_limit(1024);
+    ///
+    /// let (sym, tail) = BorrowedSymbol::with_tail_and_options(&mangled[..], &parse_options)
+    ///     .expect("Could not parse mangled symbol!");
+    ///
+    /// assert_eq!(tail, b" and some trailing junk");
+    ///
+    /// let demangled = sym.to_string();
+    /// assert_eq!(demangled, "space::foo(int, bool, char)");
+    /// ```
+    pub fn with_tail_and_options(
+        input: &'a T,
+        options: &ParseOptions,
+    ) -> Result<(BorrowedSymbol<'a>, &'a [u8])> {
         let mut substitutions = subs::SubstitutionTable::new();
 
-        let ctx = ParseContext::default();
+        let ctx = ParseContext::new(*options);
         let idx_str = IndexStr::new(input.as_ref());
         let (parsed, tail) = ast::MangledName::parse(&ctx, &mut substitutions, idx_str)?;
         debug_assert!(ctx.recursion_level() == 0);